% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDeltaLE.R
\name{getDeltaLE}
\alias{getDeltaLE}
\title{Given a country, return the increase in years of life expectancy between 1952 and 2007.}
\usage{
getDeltaLE(gap_country)
}
\arguments{
\item{gap_country}{charachter}
}
\value{
numeric vector of lenght 1
}
\description{
Given a country, return the increase in years of life expectancy between 1952 and 2007.
}
