% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDeltaLE.R
\name{getDeltaLE}
\alias{getDeltaLE}
\title{Given a country from the gapminder dataset, return the increase in years of life expectancy between 1952 and 2007.}
\usage{
getDeltaLE(gap_country)
}
\arguments{
\item{x}{factor}
}
\value{
A vector of length 1
}
\description{
Given a country from the gapminder dataset, return the increase in years of life expectancy between 1952 and 2007.
}
\examples{
getDeltaLE(France)
}
